REFACTOR : 뷰 연산 프로퍼티로 분리
